Object Pool System
Object Pooling is essential when we want to Instantiate many objects and want to maintain the performance of the devices. This comes in handy when it comes to developing games for low-end devices such as mobile.

Here is a short description of how object pooling works:

1. We create various pools for each object and pre-populate the pool instead of creating the object in an instance when it is required.
2. These pools are stored in a dictionary with appropriate tags to access the pools (usually tags are the name of the pool for eg: Enemy Pool, Bullet Pool)
3. Inside the dictionary the pools are stored in the form of QUEUE with comes with FIFO insertion & deletion this makes it handy when we want to enqueue or dequeue objects.
4. Each time any object is required it is retrieved from the pool dequeued stored in a var, set active, set transform & rotation, and then push at the back of the queue.

Also: the pool size can be expended in any movement if subsequently extra objects are required at the movement.